pyproject.toml
[project]
name = "arena-hackathon"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "matplotlib>=3.10.6",
    "numpy>=2.3.3",
    "pandas>=2.3.2",
    "python-dotenv>=1.1.1",
    "transformers>=4.56.1",
    "openai>=1.87.0,<=1.90.0",
    "pydantic>=2.11.7",
    "loguru>=0.7.3",
    "scipy>=1.16.0",
    "tokenizers>=0.22.0",
    "torch>=2.7.1",
    "tiktoken>=0.11.0",
]
